+++ src/lib/installer/crate_installer_test.rs
@@ -13,6 +13,7 @@ fn invoke_rustup_install_none() {
         rustup_component_name: None,
         min_version: None,
         version: None,
+        install_command: None,
     };
 
     let output = invoke_rustup_install(&None, &info);
@@ -30,6 +31,7 @@ fn invoke_rustup_install_fail() {
         rustup_component_name: Some("unknown_rustup_component_test".to_string()),
         min_version: None,
         version: None,
+        install_command: None,
     };
 
     let output = invoke_rustup_install(&None, &info);
@@ -49,6 +51,7 @@ fn invoke_rustup_install_with_toolchain_none() {
         rustup_component_name: None,
         min_version: None,
         version: None,
+        install_command: None,
     };
 
     let output = invoke_rustup_install(&Some(toolchain), &info);
@@ -68,6 +71,7 @@ fn invoke_rustup_install_with_toolchain_fail() {
         rustup_component_name: Some("unknown_rustup_component_test".to_string()),
         min_version: None,
         version: None,
+        install_command: None,
     };
 
     let output = invoke_rustup_install(&Some(toolchain), &info);
@@ -85,6 +89,7 @@ fn invoke_cargo_install_test() {
         rustup_component_name: Some("unknown_rustup_component_test".to_string()),
         min_version: None,
         version: None,
+        install_command: None,
     };
 
     invoke_cargo_install(&None, &info, &None, false);
@@ -103,6 +108,7 @@ fn invoke_cargo_install_with_toolchain_test() {
         rustup_component_name: Some("unknown_rustup_component_test".to_string()),
         min_version: None,
         version: None,
+        install_command: None,
     };
 
     invoke_cargo_install(&Some(toolchain), &info, &None, false);
@@ -119,6 +125,7 @@ fn install_test_test() {
         rustup_component_name: Some("unknown_rustup_component_test".to_string()),
         min_version: None,
         version: None,
+        install_command: None,
     };
 
     install(&None, &info, &None, false);
@@ -137,6 +144,7 @@ fn install_test_with_toolchain_test() {
         rustup_component_name: Some("unknown_rustup_component_test".to_string()),
         min_version: None,
         version: None,
+        install_command: None,
     };
 
     install(&Some(toolchain), &info, &None, false);
@@ -153,6 +161,7 @@ fn install_already_installed_crate_only() {
         rustup_component_name: None,
         min_version: None,
         version: None,
+        install_command: None,
     };
 
     install(&None, &info, &None, false);
@@ -160,7 +169,7 @@ fn install_already_installed_crate_only() {
 
 #[test]
 fn install_already_installed_crate_only_min_version_equal() {
-    let version = crate_version_check::get_crate_version("cargo-make").unwrap();
+    let version = crate_version_check::get_crate_version("cargo-make", None).unwrap();
     let mut version_string = String::new();
     version_string.push_str(&version.major.to_string());
     version_string.push_str(".");
@@ -177,6 +186,7 @@ fn install_already_installed_crate_only_min_version_equal() {
         rustup_component_name: None,
         min_version: Some(version_string),
         version: None,
+        install_command: None,
     };
 
     install(&None, &info, &None, false);
@@ -184,7 +194,7 @@ fn install_already_installed_crate_only_min_version_equal() {
 
 #[test]
 fn install_already_installed_crate_only_min_version_smaller() {
-    let mut version = crate_version_check::get_crate_version("cargo-make").unwrap();
+    let mut version = crate_version_check::get_crate_version("cargo-make", None).unwrap();
     if version.patch > 0 {
         version.patch = version.patch - 1;
     } else if version.minor > 0 {
@@ -209,6 +219,7 @@ fn install_already_installed_crate_only_min_version_smaller() {
         rustup_component_name: None,
         min_version: Some(version_string),
         version: None,
+        install_command: None,
     };
 
     install(&None, &info, &None, false);
@@ -216,7 +227,7 @@ fn install_already_installed_crate_only_min_version_smaller() {
 
 #[test]
 fn install_already_installed_crate_only_version_equal() {
-    let version = crate_version_check::get_crate_version("cargo-make").unwrap();
+    let version = crate_version_check::get_crate_version("cargo-make", None).unwrap();
     let mut version_string = String::new();
     version_string.push_str(&version.major.to_string());
     version_string.push_str(".");
@@ -233,6 +244,7 @@ fn install_already_installed_crate_only_version_equal() {
         rustup_component_name: None,
         min_version: None,
         version: Some(version_string),
+        install_command: None,
     };
 
     install(&None, &info, &None, false);
@@ -249,6 +261,7 @@ fn is_crate_only_info_with_rustup_component_name() {
         rustup_component_name: Some("component".to_string()),
         min_version: None,
         version: None,
+        install_command: None,
     };
 
     let crate_only_info = is_crate_only_info(&info);
@@ -267,6 +280,7 @@ fn is_crate_only_info_without_rustup_component_name() {
         rustup_component_name: None,
         min_version: None,
         version: None,
+        install_command: None,
     };
 
     let crate_only_info = is_crate_only_info(&info);